Separated IAM in Users, Stores, Access resources; Removing Manager role.

+++ SEnTri/IAMProvider/src/main/java/pt/fct/nova/id/srv/presentation/controllers/UsersController.java
@@ -1,6 +1,9 @@
 package pt.fct.nova.id.srv.presentation.controllers;
 
-import jakarta.ws.rs.core.*;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.core.Cookie;
+import jakarta.ws.rs.core.NewCookie;
+import jakarta.ws.rs.core.Response;
 import org.apache.commons.codec.binary.Base64;
 import pt.fct.nova.id.srv.application.AccessRequest;
 import pt.fct.nova.id.srv.application.IAMStore;
@@ -9,31 +12,32 @@ import pt.fct.nova.id.srv.application.clients.LocksClient;
 import pt.fct.nova.id.srv.application.clients.exception.TooManyLockRetriesException;
 import pt.fct.nova.id.srv.application.crypto.PasswordUtils;
 import pt.fct.nova.id.srv.presentation.Utils;
-import pt.fct.nova.id.srv.presentation.api.IdentityAndAccessManagementAPI;
+import pt.fct.nova.id.srv.presentation.api.UsersAPI;
 import pt.fct.nova.id.srv.presentation.api.dtos.*;
-import pt.fct.nova.id.srv.presentation.exceptions.*;
+import pt.fct.nova.id.srv.presentation.exceptions.InvalidPasswordException;
+import pt.fct.nova.id.srv.presentation.exceptions.SessionException;
+import pt.fct.nova.id.srv.presentation.exceptions.UnknownUserException;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
-import java.util.*;
 
 import static jakarta.ws.rs.core.Response.Status.*;
-import static pt.fct.nova.id.srv.application.IAMStore.*;
-import static pt.fct.nova.id.srv.presentation.Utils.extractAccessToken;
-import static pt.fct.nova.id.srv.presentation.Utils.handleSessionException;
+import static jakarta.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+import static pt.fct.nova.id.srv.presentation.Utils.*;
 import static pt.fct.nova.id.srv.presentation.api.dtos.Role.*;
+import static pt.fct.nova.id.srv.presentation.api.dtos.Role.ADMIN;
+import static pt.fct.nova.id.srv.presentation.controllers.AccessController.INSUFFICIENT_PERMISSIONS;
+import static pt.fct.nova.id.srv.presentation.controllers.StoresController.UNKNOWN_STORE;
 
-public class IAMController implements IdentityAndAccessManagementAPI {
+@Path("/users")
+public class UsersController implements UsersAPI {
     private static final String SUCCESSFUL_AUTH = "Successful authentication.";
-    private static final String UNKNOWN_USER = "User not found.";
     private static final String INVALID_PASSWORD = "Password mismatch.";
-    private static final String INTERNAL_ERROR = "Internal error.";
+    public static final String UNKNOWN_USER = "User not found.";
     private static final String USER_ALREADY_EXISTS = "User already exists.";
+    private static final String CAN_NOT_DELETE_STORE_OWNER = "Can not delete store owners.";
     private static final String SUCCESSFUL_USER_REGISTER = "Successful user registration.";
     private static final String SUCCESSFUL_USER_DELETE = "Successful user deletion.";
-    private static final String UNKNOWN_STORE = "Store not found.";
-    private static final String INSUFFICIENT_PERMISSIONS = "Insufficient permissions to execute request.";
-    private static final String OPERATION_TIMEOUT = "Operation timeout.";
     private static final String SUCCESSFUL_ACCESS_REVOCATION = "Successful access revocation.";
     private static final String SUCCESSFUL_ACCESS_REQUEST_ISSUED = "Successful issued access request.";
     private static final String SUCCESSFUL_ACCESS_GRANT = "Successful access grant.";
@@ -41,23 +45,13 @@ public class IAMController implements IdentityAndAccessManagementAPI {
     private static final String SUCCESSFUL_ROLE_GRANT = "Successful role grant.";
     private static final String REQUEST_NOT_FOUND = "Request not found.";
     private static final String SUCCESSFUL_REQUEST_PROCESSING = "Successful request processing.";
-    private static final String STORE_ALREADY_EXISTS = "Store already exists.";
-    private static final String SUCCESSFUL_STORE_ACCESS_POLICY_CREATION = "Successful creation of store access policy.";
-    private static final String SUCCESSFUL_STORE_ACCESS_POLICY_DELETION = "Successful deletion of store access policy.";
-    private static final String ACCESS_ALLOWED = "Access allowed.";
-    public static final String ACCESS_FORBIDDEN = "Access forbidden.";
-    public static final String NO_ACCESS_TOKEN = "Malformed request: bearer token required.";
-    public static final String UNKNOWN_OR_EXPIRED_TOKEN = "Token not found or expired.";
-    private static final String SUCCESSFUL_ACCESS_TOKEN_DELETION = "Successful deletion of access token.";
-
     private static final String REQUEST_DECISION_MALFORMED = "Request decision malformed.";
-
 
     @Override
     public Response auth(AuthForm credentials) {
         try {
             String username = credentials.getUsername();
-            checkPassword(username, credentials.getPassword());
+            Utils.checkPassword(username, credentials.getPassword());
             NewCookie cookie = IAMStore.cacheSession(credentials.getUsername());
             return Response.ok(SUCCESSFUL_AUTH).cookie(cookie).build();
         } catch (UnknownUserException e) {
@@ -67,14 +61,6 @@ public class IAMController implements IdentityAndAccessManagementAPI {
         } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
             return Response.ok(INTERNAL_ERROR).status(Response.Status.INTERNAL_SERVER_ERROR).build();
         }
-    }
-
-    private void checkPassword(String username, String password) throws InvalidPasswordException, UnknownUserException, NoSuchAlgorithmException, InvalidKeySpecException {
-        byte[] hash = Base64.decodeBase64(IAMStore.getPassword(username));
-        if (hash == null)
-            throw new UnknownUserException();
-        if (!PasswordUtils.verify(password, hash))
-            throw new InvalidPasswordException();
     }
 
     @Override
@@ -103,6 +89,11 @@ public class IAMController implements IdentityAndAccessManagementAPI {
             Utils.authCheck(cookie, username);
             String lockID = LocksClient.acquireUserLock(username);
             IAMStore.deleteUser(username);
+            //TODO: Check that doesn't own any store
+            if(IAMStore.checkIfOwnsAny(username)) {
+                LocksClient.releaseUserLock(username, lockID);
+                return Response.ok(CAN_NOT_DELETE_STORE_OWNER).build();
+            }
             LocksClient.deleteAllUserLocks(username);
             LocksClient.releaseUserLock(username, lockID);
             return Response.ok(SUCCESSFUL_USER_DELETE).build();
@@ -338,258 +329,6 @@ public class IAMController implements IdentityAndAccessManagementAPI {
             return Response.ok(OPERATION_TIMEOUT).status(INTERNAL_SERVER_ERROR).build();
         } catch (InterruptedException e) {
             return Response.ok(INTERNAL_ERROR).status(INTERNAL_SERVER_ERROR).build();
-        }
-    }
-
-    @Override
-    public Response createStoreAccessPolicy(Cookie cookie, String username, StoreForm form) {
-        try {
-            String issuer = form.getIssuer();
-            Utils.authCheck(cookie, issuer);
-            Role issuerRole = IAMStore.getRole(issuer);
-
-            if (issuer.equals(username) && issuerRole.equals(BASIC))
-                return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-            if (!issuer.equals(username) && issuerRole.equals(BASIC) || issuerRole.equals(PRIVILEGED))
-                return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-            if (!IAMStore.userExists(username))
-                return Response.ok(UNKNOWN_USER).status(NOT_FOUND).build();
-            if (IAMStore.getRole(username).equals(BASIC))
-                return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-
-            String storeID = form.getStoreID();
-            String lockID = LocksClient.acquireStoreLock(issuer, storeID);
-            if (!IAMStore.storeAccessPolicyExists(storeID)) {
-                LocksClient.releaseStoreLock(issuer, storeID, lockID);
-                return Response.ok(STORE_ALREADY_EXISTS).status(BAD_REQUEST).build();
-            }
-            IAMStore.saveStoreAccessPolicy(storeID, username, new HashSet<>(), new HashSet<>());
-            LocksClient.releaseStoreLock(issuer, storeID, lockID);
-            return Response.ok(SUCCESSFUL_STORE_ACCESS_POLICY_CREATION).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        } catch (TooManyLockRetriesException e) {
-            return Response.ok(OPERATION_TIMEOUT).status(INTERNAL_SERVER_ERROR).build();
-        } catch (InterruptedException e) {
-            return Response.ok(INTERNAL_ERROR).status(Response.Status.INTERNAL_SERVER_ERROR).build();
-        }
-    }
-
-    @Override
-    public Response deleteStoreAccessPolicy(Cookie cookie, String storeID, List<String> authorizationHeaders) {
-        try {
-            String tokenID = extractAccessToken(authorizationHeaders);
-            if (tokenID == null)
-                return Response.ok(NO_ACCESS_TOKEN).status(BAD_REQUEST).build();
-
-            Map<String, String> token = IAMStore.getToken(tokenID);
-            if (token == null || token.isEmpty())
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            String username = token.get(TOKEN_USER_FIELD);
-            Utils.authCheck(cookie, username);
-
-            String tokenStoreID = Objects.requireNonNull(token.get(TOKEN_STORE_FIELD));
-
-            if (!IAMStore.storeAccessPolicyExists(storeID)) {
-                if (storeID.equals(tokenStoreID))
-                    IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(UNKNOWN_STORE).status(NOT_FOUND).build();
-            }
-
-            if (!tokenStoreID.equals(storeID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            Role role = IAMStore.getRole(username);
-
-            if (role.equals(BASIC) || (role.equals(PRIVILEGED) && !IAMStore.checkIfOwns(username, storeID)))
-                return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-
-            String lockID = token.get(TOKEN_LOCK_FIELD);
-            if (lockID == null)
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            else if (!LocksClient.checkIfStoreLockExists(storeID, lockID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            IAMStore.deleteStoreAccessPolicy(storeID);
-            LocksClient.releaseStoreLock(username, storeID, lockID);
-            return Response.ok(SUCCESSFUL_STORE_ACCESS_POLICY_DELETION).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        }
-    }
-
-    @Override
-    public Response getReadAccess(Cookie cookie, String storeID, List<String> authorizationHeaders) {
-        try {
-            String tokenID = extractAccessToken(authorizationHeaders);
-            if (tokenID == null)
-                return Response.ok(NO_ACCESS_TOKEN).status(BAD_REQUEST).build();
-
-            Map<String, String> token = IAMStore.getToken(tokenID);
-            if (token == null || token.isEmpty())
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            String username = token.get(TOKEN_USER_FIELD);
-            Utils.authCheck(cookie, username);
-
-            String tokenStoreID = Objects.requireNonNull(token.get(TOKEN_STORE_FIELD));
-
-            if (!IAMStore.storeAccessPolicyExists(storeID)) {
-                if (storeID.equals(tokenStoreID))
-                    IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(UNKNOWN_STORE).status(NOT_FOUND).build();
-            }
-
-            if (!tokenStoreID.equals(storeID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            if (!IAMStore.checkIfUserHasReadAccess(username, storeID) &&
-                    !IAMStore.checkIfUserHasWriteAccess(username, storeID) &&
-                    !IAMStore.getRole(username).equals(ADMIN)) {
-                IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            }
-            return Response.ok(ACCESS_ALLOWED).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        }
-    }
-
-
-    @Override
-    public Response getWriteAccess(Cookie cookie, String storeID, List<String> authorizationHeaders) {
-        try {
-            String tokenID = extractAccessToken(authorizationHeaders);
-            if (tokenID == null)
-                return Response.ok(NO_ACCESS_TOKEN).status(BAD_REQUEST).build();
-
-            Map<String, String> token = IAMStore.getToken(tokenID);
-            if (token == null || token.isEmpty())
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            String username = token.get(TOKEN_USER_FIELD);
-            Utils.authCheck(cookie, username);
-
-            String tokenStoreID = Objects.requireNonNull(token.get(TOKEN_STORE_FIELD));
-
-            if (!IAMStore.storeAccessPolicyExists(storeID)) {
-                if (storeID.equals(tokenStoreID))
-                    IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(UNKNOWN_STORE).status(NOT_FOUND).build();
-            }
-
-            if (!tokenStoreID.equals(storeID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            String lockID = token.get(TOKEN_LOCK_FIELD);
-            if (lockID == null)
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            else if (!LocksClient.checkIfStoreLockExists(storeID, lockID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            if (!IAMStore.checkIfUserHasWriteAccess(username, storeID) && !IAMStore.getRole(username).equals(ADMIN)) {
-                IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            }
-            return Response.ok(ACCESS_ALLOWED).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        }
-    }
-
-    @Override
-    public Response getOwnerAccess(Cookie cookie, String storeID, List<String> authorizationHeaders) {
-        try {
-            String tokenID = extractAccessToken(authorizationHeaders);
-            if (tokenID == null)
-                return Response.ok(NO_ACCESS_TOKEN).status(BAD_REQUEST).build();
-
-            Map<String, String> token = IAMStore.getToken(tokenID);
-            if (token == null || token.isEmpty())
-                return Response.ok(UNKNOWN_OR_EXPIRED_TOKEN).status(NOT_FOUND).build();
-
-            String username = token.get(TOKEN_USER_FIELD);
-            Utils.authCheck(cookie, username);
-            String tokenStoreID = Objects.requireNonNull(token.get(TOKEN_STORE_FIELD));
-
-            if (!IAMStore.storeAccessPolicyExists(storeID)) {
-                if (storeID.equals(tokenStoreID))
-                    IAMStore.deleteAccessToken(tokenID, token);
-                return Response.ok(UNKNOWN_STORE).status(NOT_FOUND).build();
-            }
-
-            if (!tokenStoreID.equals(storeID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            String lockID = token.get(TOKEN_LOCK_FIELD);
-            if (lockID == null)
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            else if (!LocksClient.checkIfStoreLockExists(storeID, lockID))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-
-            if (!IAMStore.checkIfOwns(username, storeID) && !IAMStore.getRole(username).equals(ADMIN))
-                return Response.ok(ACCESS_FORBIDDEN).status(FORBIDDEN).build();
-            return Response.ok(ACCESS_ALLOWED).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        }
-    }
-
-    @Override
-    public Response createAccessToken(Cookie cookie, String username, StoreForm form) {
-        try {
-            String issuer = form.getIssuer();
-            String storeID = form.getStoreID();
-            Utils.authCheck(cookie, issuer);
-            if (!IAMStore.userExists(username))
-                return Response.ok(UNKNOWN_USER).status(BAD_REQUEST).build();
-            if (!IAMStore.storeAccessPolicyExists(storeID))
-                return Response.ok(UNKNOWN_STORE).status(BAD_REQUEST).build();
-
-            Role issuerRole = IAMStore.getRole(issuer);
-            if (!issuer.equals(username)) {
-                if (issuerRole.equals(BASIC) || (issuerRole.equals(PRIVILEGED) && !IAMStore.checkIfOwns(issuer, storeID)))
-                    return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-            }
-            Role role = IAMStore.getRole(issuer);
-
-            if (!role.equals(ADMIN) &&
-                    !IAMStore.checkIfUserHasReadAccess(username, storeID) &&
-                    !IAMStore.checkIfUserHasWriteAccess(username, storeID) &&
-                    !IAMStore.checkIfOwns(username, storeID))
-                return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-
-            return Response.ok(IAMStore.saveToken(username, storeID)).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
-        }
-    }
-
-    @Override
-    public Response deleteAccessToken(Cookie cookie, String username, List<String> authorizationHeaders) {
-        try {
-            String tokenID = extractAccessToken(authorizationHeaders);
-            if (tokenID == null)
-                return Response.ok(NO_ACCESS_TOKEN).status(BAD_REQUEST).build();
-
-            Map<String, String> token = IAMStore.getToken(tokenID);
-            if (token == null || token.isEmpty())
-                return Response.ok(UNKNOWN_OR_EXPIRED_TOKEN).status(NOT_FOUND).build();
-
-            String tokenOwner = token.get(TOKEN_USER_FIELD);
-            Utils.authCheck(cookie, username);
-            Role role = IAMStore.getRole(username);
-            String tokenStoreID = Objects.requireNonNull(token.get(TOKEN_STORE_FIELD));
-
-            if (!username.equals(tokenOwner)) {
-                if (role.equals(BASIC) || (role.equals(PRIVILEGED) && !IAMStore.checkIfOwns(username, tokenStoreID)))
-                    return Response.ok(INSUFFICIENT_PERMISSIONS).status(FORBIDDEN).build();
-            }
-            IAMStore.deleteAccessToken(tokenID, token);
-            return Response.ok(SUCCESSFUL_ACCESS_TOKEN_DELETION).build();
-        } catch (SessionException e) {
-            return handleSessionException(e);
         }
     }
 }
